modules/cdi-commons/src/main/java/org/rapidpm/commons/cdi/logger/LoggerProducer.java
/*
 * Copyright [2013] [www.rapidpm.org / Sven Ruppert (sven.ruppert@rapidpm.org)]
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

package org.rapidpm.commons.cdi.logger;

import javax.enterprise.inject.Produces;
import javax.enterprise.inject.spi.BeanManager;
import javax.enterprise.inject.spi.InjectionPoint;

import org.rapidpm.commons.cdi.logger.Logger;


/**
 * User: Sven Ruppert
 * Date: 03.06.13
 * Time: 09:04
 */
public class LoggerProducer {

    /**
     * @param injectionPoint
     * @return logger
     */
    @Produces
    @CDILogger
    public Logger produceLog4JLogger(InjectionPoint injectionPoint, BeanManager beanManager) {
        final Class<?> declaringClass = injectionPoint.getMember().getDeclaringClass();
        return new Logger(declaringClass);
    }
}
